tweak: improve recognizability of auto-save files
Make the resulting auto-save files easier to recognize when perused by
hand (in case of a fatal data loss and recovery situation), by prefixing
the hashed autosave file with the target file name.

Also updates the explanatory comment above it to better explain the
purpose of these defaults.

Also uses file-name-concat in case auto-save-list-file-prefix doesn't
end in a slash.